adding plugins
