+++ internal/cmd/build.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"gobuffalo/cli/internal/genny/build"
+	"github.com/gobuffalo/cli/internal/genny/build"
 
 	"github.com/gobuffalo/genny/v2"
 	"github.com/gobuffalo/logger"
