+++ internal/cmd/generate/mailer.go
@@ -3,7 +3,7 @@ package generate
 import (
 	"context"
 
-	"gobuffalo/cli/internal/genny/mail"
+	"github.com/gobuffalo/cli/internal/genny/mail"
 
 	"github.com/gobuffalo/flect/name"
 	"github.com/gobuffalo/genny/v2"
