+++ internal/cmd/generate/task.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"os"
 
-	"gobuffalo/cli/internal/genny/grift"
+	"github.com/gobuffalo/cli/internal/genny/grift"
 
 	"github.com/gobuffalo/genny/v2"
 	"github.com/gobuffalo/genny/v2/gogen"
