+++ internal/cmd/new.go
@@ -11,15 +11,15 @@ import (
 	"path/filepath"
 	"strings"
 
-	"gobuffalo/cli/internal/genny/assets/standard"
-	"gobuffalo/cli/internal/genny/assets/webpack"
-	"gobuffalo/cli/internal/genny/ci"
-	"gobuffalo/cli/internal/genny/docker"
-	"gobuffalo/cli/internal/genny/newapp/api"
-	"gobuffalo/cli/internal/genny/newapp/core"
-	"gobuffalo/cli/internal/genny/newapp/web"
-	"gobuffalo/cli/internal/genny/refresh"
-	"gobuffalo/cli/internal/genny/vcs"
+	"github.com/gobuffalo/cli/internal/genny/assets/standard"
+	"github.com/gobuffalo/cli/internal/genny/assets/webpack"
+	"github.com/gobuffalo/cli/internal/genny/ci"
+	"github.com/gobuffalo/cli/internal/genny/docker"
+	"github.com/gobuffalo/cli/internal/genny/newapp/api"
+	"github.com/gobuffalo/cli/internal/genny/newapp/core"
+	"github.com/gobuffalo/cli/internal/genny/newapp/web"
+	"github.com/gobuffalo/cli/internal/genny/refresh"
+	"github.com/gobuffalo/cli/internal/genny/vcs"
 
 	pop "github.com/gobuffalo/buffalo-pop/v2/genny/newapp"
 	"github.com/gobuffalo/cli/internal/takeon/github.com/markbates/errx"
