+++ internal/cmd/plugins/add.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"strings"
 
-	"gobuffalo/cli/internal/genny/add"
+	"github.com/gobuffalo/cli/internal/genny/add"
 
 	"github.com/gobuffalo/cli/internal/plugins/plugdeps"
 	"github.com/gobuffalo/cli/internal/takeon/github.com/markbates/errx"
