+++ internal/genny/actions/actions.go
@@ -10,7 +10,7 @@ import (
 	"github.com/gobuffalo/packr/v2"
 )
 
-var box = packr.New("gobuffalo/cli/internal/genny/actions/templates", "../actions/templates")
+var box = packr.New("github.com/gobuffalo/cli/internal/genny/actions/templates", "../actions/templates")
 
 // New returns a new generator for build actions on a Buffalo app
 func New(opts *Options) (*genny.Generator, error) {
