+++ internal/genny/newapp/core/core_test.go
@@ -3,7 +3,7 @@ package core
 import (
 	"testing"
 
-	"gobuffalo/cli/internal/genny/docker"
+	"github.com/gobuffalo/cli/internal/genny/docker"
 
 	"github.com/gobuffalo/envy"
 	"github.com/gobuffalo/genny/v2/gentest"
