+++ internal/genny/newapp/web/web.go
@@ -3,9 +3,9 @@ package web
 import (
 	"html/template"
 
-	"gobuffalo/cli/internal/genny/assets/standard"
-	"gobuffalo/cli/internal/genny/assets/webpack"
-	"gobuffalo/cli/internal/genny/newapp/core"
+	"github.com/gobuffalo/cli/internal/genny/assets/standard"
+	"github.com/gobuffalo/cli/internal/genny/assets/webpack"
+	"github.com/gobuffalo/cli/internal/genny/newapp/core"
 
 	"github.com/gobuffalo/genny/v2"
 	"github.com/gobuffalo/genny/v2/gogen"
