+++ internal/genny/newapp/web/web_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"gobuffalo/cli/internal/genny/newapp/core"
+	"github.com/gobuffalo/cli/internal/genny/newapp/core"
 
 	"github.com/gobuffalo/envy"
 	"github.com/gobuffalo/genny/v2/gentest"
