+++ internal/genny/plugins/install/install.go
@@ -7,7 +7,7 @@ import (
 	"os/exec"
 	"path/filepath"
 
-	"gobuffalo/cli/internal/genny/add"
+	"github.com/gobuffalo/cli/internal/genny/add"
 
 	"github.com/gobuffalo/cli/internal/plugins/plugdeps"
 	"github.com/gobuffalo/genny/v2"
