+++ internal/plugins/plugdeps/plugdeps_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gobuffalo/buffalo/internal/takeon/github.com/markbates/errx"
+	"github.com/gobuffalo/cli/internal/takeon/github.com/markbates/errx"
 	"github.com/gobuffalo/meta"
 	"github.com/stretchr/testify/require"
 )
