+++ internal/plugins/plugins.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gobuffalo/buffalo/internal/takeon/github.com/markbates/errx"
 	"github.com/gobuffalo/cli/internal/plugins/plugdeps"
+	"github.com/gobuffalo/cli/internal/takeon/github.com/markbates/errx"
 	"github.com/gobuffalo/envy"
 	"github.com/gobuffalo/meta"
 	"github.com/karrick/godirwalk"
